refactor(behavior_path_planner): separate drivable area functions
