+++ planning/behavior_path_planner/src/utils/goal_planner/shift_pull_over.cpp
@@ -14,6 +14,7 @@
 
 #include "behavior_path_planner/utils/goal_planner/shift_pull_over.hpp"
 
+#include "behavior_path_planner/utils/drivable_area_expansion/static_drivable_area.hpp"
 #include "behavior_path_planner/utils/goal_planner/util.hpp"
 #include "behavior_path_planner/utils/path_utils.hpp"
